update authtenticate, author, edit user

+++ routes/web.php
@@ -1,5 +1,7 @@
 <?php
 
+use App\Http\Controllers\AgentTravelController;
+use App\Http\Controllers\GuestController;
 use App\Http\Controllers\LoginController;
 use App\Http\Controllers\RegisterController;
 use App\Http\Controllers\UserController;
@@ -15,17 +17,25 @@ use Illuminate\Support\Facades\Route;
 | contains the "web" middleware group. Now create something great!
 |
 */
-// AUTH
-Route::get('/login', [LoginController::class, 'index']);
-Route::get('/register', [RegisterController::class, 'index']);
-Route::post('/register', [RegisterController::class, 'create']);
+
+Route::post('/logout', [LoginController::class, 'logout']);
+
+// GUEST ROUTES
+Route::get('/', [GuestController::class, 'index'])->middleware('guest');;
+Route::get('/login', [LoginController::class, 'index'])->name('login')->middleware('guest');
+Route::get('/register', [RegisterController::class, 'index'])->middleware('guest');
+Route::post('/login', [LoginController::class, 'authenticate'])->middleware('guest');
+Route::post('/register', [RegisterController::class, 'create'])->middleware('guest');
 
 // ADMIN ROUTES
-Route::get('/', [UserController::class, 'index']);
-Route::get('/users/{user}', [UserController::class, 'show']);
-Route::get('/users/edit/{user}', [UserController::class, 'edit']);
-Route::delete('/users/delete/{user}', [UserController::class, 'destroy']);
-Route::put('/users/update/{user}', [UserController::class, 'update']);
+Route::get('/admin', [UserController::class, 'index'])->middleware('admin');
+Route::get('/admin/users/{user}', [UserController::class, 'show'])->middleware('admin');
+Route::get('/admin/users/edit/{user}', [UserController::class, 'edit'])->middleware('admin');
+Route::delete('/admin/users/delete/{user}', [UserController::class, 'destroy'])->middleware('admin');
+Route::put('/admin/users/update/{user}', [UserController::class, 'update'])->middleware('admin');
+
+// AGENT TRAVEL ROUTES
+Route::get('/agent-travel', [AgentTravelController::class, 'index'])->middleware('agent_travel');
